requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
gspread==5.7.2
google-auth==2.20.0
pdfplumber==0.9.0
python-docx==0.8.11
pytest==7.4.0
twilio==9.8.4
langchain>=0.0.300
google-cloud-aiplatform>=1.26.0
jsonschema>=4.0.0
python-dateutil>=2.8.2
langchain
